consoledate most of the stuff in seperate containers
