+++ who-metrics-ui/src/components/DashboardExample.tsx
@@ -10,31 +10,17 @@ import {
   TabGroup,
   TabPanel,
   TabPanels,
-  BadgeDelta,
   DeltaType,
-  Flex,
-  Metric,
-  ProgressBar,
-  MultiSelect,
-  MultiSelectItem,
-  Select,
-  SelectItem,
-  Table,
-  TableBody,
-  TableCell,
-  TableHead,
-  TableHeaderCell,
-  TableRow,
 } from "@tremor/react";
 
-import { InfoIcon } from "@primer/octicons-react";
-import { Box, Tooltip } from "@primer/react";
+import { Box } from "@primer/react";
 import Image from "next/image";
 // eslint-disable-next-line import/no-unresolved, import/extensions
 import logo from "@/images/who-logo-wide.svg";
 import { ChartView } from "./";
 
-import { useState } from "react";
+import KpiCard from "./KpiCard";
+import { PerformanceHistoryTable } from "./PerformanceHistoryTable";
 
 type Kpi = {
   title: string;
@@ -106,78 +92,7 @@ export const performance: DailyPerformance[] = [
   },
 ];
 
-export type SalesPerson = {
-  name: string;
-  leads: number;
-  sales: string;
-  quota: string;
-  variance: string;
-  region: string;
-  status: string;
-};
-
-export const salesPeople: SalesPerson[] = [
-  {
-    name: "Peter Doe",
-    leads: 45,
-    sales: "1,000,000",
-    quota: "1,200,000",
-    variance: "low",
-    region: "Region A",
-    status: "overperforming",
-  },
-  {
-    name: "Lena Whitehouse",
-    leads: 35,
-    sales: "900,000",
-    quota: "1,000,000",
-    variance: "low",
-    region: "Region B",
-    status: "average",
-  },
-  {
-    name: "Phil Less",
-    leads: 52,
-    sales: "930,000",
-    quota: "1,000,000",
-    variance: "medium",
-    region: "Region C",
-    status: "underperforming",
-  },
-  {
-    name: "John Camper",
-    leads: 22,
-    sales: "390,000",
-    quota: "250,000",
-    variance: "low",
-    region: "Region A",
-    status: "overperforming",
-  },
-  {
-    name: "Max Balmoore",
-    leads: 49,
-    sales: "860,000",
-    quota: "750,000",
-    variance: "low",
-    region: "Region B",
-    status: "overperforming",
-  },
-];
-
-const deltaTypes: { [key: string]: DeltaType } = {
-  average: "unchanged",
-  overperforming: "moderateIncrease",
-  underperforming: "moderateDecrease",
-};
-
 export const DashboardExample = () => {
-  const [selectedStatus, setSelectedStatus] = useState("all");
-  const [selectedNames, setSelectedNames] = useState<string[]>([]);
-
-  const isSalesPersonSelected = (salesPerson: SalesPerson) =>
-    (salesPerson.status === selectedStatus || selectedStatus === "all") &&
-    (selectedNames.includes(salesPerson.name) || selectedNames.length === 0);
-
   return (
     <main className="px-18 py-18">
       <Box
@@ -203,22 +118,15 @@ export const DashboardExample = () => {
           <TabPanel>
             <Grid numItemsLg={3} className="mt-6 gap-6">
               {kpiData.map((item) => (
-                <Card key={item.title}>
-                  <Flex alignItems="start">
-                    <div className="truncate">
-                      <Text>{item.title}</Text>
-                      <Metric className="truncate">{item.metric}</Metric>
-                    </div>
-                    <BadgeDelta deltaType={item.deltaType}>
-                      {item.delta}
-                    </BadgeDelta>
-                  </Flex>
-                  <Flex className="mt-4 space-x-2">
-                    <Text className="truncate">{`${item.progress}% (${item.metric})`}</Text>
-                    <Text>{item.target}</Text>
-                  </Flex>
-                  <ProgressBar value={item.progress} className="mt-2" />
-                </Card>
+                <KpiCard
+                  key={item.title}
+                  title={item.title}
+                  metric={item.metric}
+                  progress={item.progress}
+                  target={item.target}
+                  delta={item.delta}
+                  deltaType={item.deltaType}
+                />
               ))}
             </Grid>
             <div className="mt-6">
@@ -228,109 +136,7 @@ export const DashboardExample = () => {
             </div>
           </TabPanel>
           <TabPanel>
-            <div className="mt-6">
-              <Card>
-                <>
-                  <div>
-                    <Flex
-                      className="space-x-0.5"
-                      justifyContent="start"
-                      alignItems="center"
-                    >
-                      <Title> Performance Hasan </Title>
-                      <Tooltip aria-label="Shows daily increase or decrease of particular domain">
-                        <InfoIcon size={24} />
-                      </Tooltip>
-                    </Flex>
-                  </div>
-                  <div className="flex space-x-2">
-                    <MultiSelect
-                      className="max-w-full sm:max-w-xs"
-                      onValueChange={setSelectedNames}
-                      placeholder="Select Salespeople..."
-                    >
-                      {salesPeople.map((item) => (
-                        <MultiSelectItem key={item.name} value={item.name}>
-                          {item.name}
-                        </MultiSelectItem>
-                      ))}
-                    </MultiSelect>
-                    <Select
-                      className="max-w-full sm:max-w-xs"
-                      defaultValue="all"
-                      onValueChange={setSelectedStatus}
-                    >
-                      <SelectItem value="all">All Performances</SelectItem>
-                      <SelectItem value="overperforming">
-                        Overperforming
-                      </SelectItem>
-                      <SelectItem value="average">Average</SelectItem>
-                      <SelectItem value="underperforming">
-                        Underperforming
-                      </SelectItem>
-                    </Select>
-                  </div>
-                  <Table className="mt-6">
-                    <TableHead>
-                      <TableRow>
-                        <TableHeaderCell>Name</TableHeaderCell>
-                        <TableHeaderCell className="text-right">
-                          Leads
-                        </TableHeaderCell>
-                        <TableHeaderCell className="text-right">
-                          Sales ($)
-                        </TableHeaderCell>
-                        <TableHeaderCell className="text-right">
-                          Quota ($)
-                        </TableHeaderCell>
-                        <TableHeaderCell className="text-right">
-                          Variance
-                        </TableHeaderCell>
-                        <TableHeaderCell className="text-right">
-                          Region
-                        </TableHeaderCell>
-                        <TableHeaderCell className="text-right">
-                          Status
-                        </TableHeaderCell>
-                      </TableRow>
-                    </TableHead>
-
-                    <TableBody>
-                      {salesPeople
-                        .filter((item) => isSalesPersonSelected(item))
-                        .map((item) => (
-                          <TableRow key={item.name}>
-                            <TableCell>{item.name}</TableCell>
-                            <TableCell className="text-right">
-                              {item.leads}
-                            </TableCell>
-                            <TableCell className="text-right">
-                              {item.sales}
-                            </TableCell>
-                            <TableCell className="text-right">
-                              {item.quota}
-                            </TableCell>
-                            <TableCell className="text-right">
-                              {item.variance}
-                            </TableCell>
-                            <TableCell className="text-right">
-                              {item.region}
-                            </TableCell>
-                            <TableCell className="text-right">
-                              <BadgeDelta
-                                deltaType={deltaTypes[item.status]}
-                                size="xs"
-                              >
-                                {item.status}
-                              </BadgeDelta>
-                            </TableCell>
-                          </TableRow>
-                        ))}
-                    </TableBody>
-                  </Table>
-                </>
-              </Card>
-            </div>
+            <PerformanceHistoryTable />
           </TabPanel>
         </TabPanels>
       </TabGroup>
